Add unit test
Since libmebo is a library that supposed to work with
third-party encoder implementations, creating a unit-test
is tricky. So to simplify the dependencies, we came up with
a fake-encoder test which generates random frame sizes
based on some real-life encode use case heuristics and feed
to libmebo.

eg: how to test:
   ./fake-test --codec=VP9 --preset=0 --framecount=120