Add BRC backend implementation selection support
Libmebo allows having multiple implementations for the
same codec. It is up to the user to select which algorithm
to use. The libmebo_rate_controller_new() API has a second
argument to select the algorithm. Or the user can stick
with the default algorithm by passing LIBMEBO_BRC_ALGORITHM_DEFAULT
as the second argument.