Fixes to satisfy static code analyzer
These are mostly trivial fixes based on klockwork static code
analysis. One major change is the increment in the max size of
the number of spatial & temporal layers. But please keep in mind
that the limit of num_ss & num_tl definitions in libmebo interface
don't mean that all the codecs can support those. The init() routine
is responsible for validating the layer limit on a per codec basis.