brc: vp9: libvpx_derived: Remove the gf_boost support
The golden frame boost support for CBR in libvpx was based on
a gf_boost input param from the user which is disabled by default.
The vp9rtc standalone brc is also not supporting this feature.
This patch removes the code blocks that are enabled based
on gf_boost enablement.