doc: Clarify under/over_shoot_pct attributes
under_shoot_pct: This factor controls the maximum amount
of bits that can be subtracted from the target bitrate
in order to compensate for prior overshoot.

over_shoot_pct: This factor controls the maximum amount
of bits that can be added to the target bitrate in order
to compensate for prior undershoot.